requirements_jn.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
alabaster==0.7.13
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
CacheControl==0.13.1
cachetools==5.3.1
catboost==1.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.1.1
coverage==7.4.1
cryptography==42.0.1
cycler==0.12.1
Cython==3.0.8
dataclasses==0.6
datamol==0.11.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
descriptastorus==2.6.1
docker-pycreds==0.4.0
docutils==0.20.1
easydict
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.0
fastjsonschema==2.19.1
filelock==3.13.0
firebase-admin==6.0.1
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
furo==2023.9.10
future==0.18.3
fuzzywuzzy==0.18.0
gcloud==0.18.3
gcsfs==2023.12.2.post1
gensim==4.3.2
# -e git+https://github.com/diamondspark/GFN-Pretrain.git@65dc445a66a9a0309a78b2b42a57f3a4afb91912#egg=gflownet
# -e git+https://github.com/Valdes-Tresanco-MS/AutoDockTools_py3
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.10.1
google-api-python-client==2.115.0
google-auth==2.23.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-firestore==2.7.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.56.4
googledrivedownloader==0.4
graphviz==0.20.1
grpcio==1.59.0
grpcio-status==1.48.2
gspread==6.0.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.18.0
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
# install==1.3.5
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==7.4.9
jupyter_core==5.4.0
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jwcrypto==1.5.1
kiwisolver==1.4.5
littleutils==0.2.2
loguru==0.7.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
meeko==0.5.1
mistune==3.0.2
mordred==1.2.0
msgpack==1.0.7
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
ogb==1.3.4
olorenchemengine==1.0.14
outdated==0.2.2
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
PubChemPy==1.0.4
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyDeprecate==0.3.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
# Pyrebase4==4.7.1
PyTDC==0.4.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-timeout==2.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-jwt==4.1.0
pytorch-lightning==1.4.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
rdflib==7.0.0
rdkit==2023.3.3
rdkit-pypi==2022.9.5
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
selfies==2.1.1
Send2Trash==1.8.2
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==59.5.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.1.2
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
StrEnum==0.4.15
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
# torch==1.10.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
# torch-cluster==1.6.0 -f https://pytorch-geometric.com/whl/torch-1.10.0+cu111.html
# torch-geometric==2.0.3
# torch-scatter==2.0.9 -f https://data.pyg.org/whl/torch-1.10.0+cu111.html
# torch-sparse==0.6.13 -f https://data.pyg.org/whl/torch-1.10.0+cu111.html 
torchmetrics==0.6.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
useful_rdkit_utils==0.2.7
wandb==0.15.12
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.16.0
xarray==2023.1.0
xgboost==1.4.2
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0
